14/02/2021 - HomeWork

+++ Entities/Concrete/Brand.cs
@@ -7,7 +7,7 @@ namespace Entities.Concrete
 {
     public class Brand:IEntity
     {
-        public int BrandId { get; set; }
+        public short BrandId { get; set; }
         public string BrandName { get; set; }
     }
 }
